Double funzionanti

+++ FitTracker/app/src/main/java/com/example/fittracker/model/Pasto.kt
@@ -10,7 +10,7 @@ data class Pasto(
     val image : String,
     val nome : String,
     val proteine : Double,
-    val quantita : Int
+    val quantita : Double
 
-) { constructor(): this(0.0,0.0,0.0,"","","",0.0,0)
+) { constructor(): this(0.0,0.0,0.0,"","","",0.0,0.0)
 }
